requirements.txt
pandas
matplotlib
SQLAlchemy
mysql-connector-python
python-dotenv
jupyter